feat: migrate topbar component back to toolkit
