+++ apps/console/.env
@@ -11,3 +11,4 @@ NEXT_PUBLIC_USAGE_COLLECTION_ENABLED=true
 NEXT_PUBLIC_SET_SECURE_COOKIE=false
 NEXT_PUBLIC_AMPLITUDE_KEY=9823fa6e3ff904bec67a8fc90db82fb9
 NEXT_PUBLIC_CONSOLE_BASE_URL=http://localhost:3000
+NEXT_PUBLIC_APP_ENV=CE
