--- apps/console/src/components/top-bar/topBarData.tsx
@@ -1,24 +1,0 @@
-import { Icons } from "@instill-ai/design-system";
-
-export const tobBarData = [
-  {
-    pathName: "pipelines",
-    Icon: Icons.Pipeline,
-    name: "Pipelines",
-  },
-  {
-    pathName: "connectors",
-    Icon: Icons.IntersectSquare,
-    name: "Connectors",
-  },
-  {
-    pathName: "models",
-    Icon: Icons.Cube01,
-    name: "Model Hub",
-  },
-  {
-    pathName: "dashboard",
-    Icon: Icons.BarChartSquare02,
-    name: "Dashboard",
-  },
-];
