+++ apps/console/src/pages/[entity]/dashboard/index.tsx
@@ -3,9 +3,10 @@ import { Logo } from "@instill-ai/design-system";
 import {
   DashboardPipelineListPageMainView,
   PageBase,
+  Topbar,
 } from "@instill-ai/toolkit";
 
-import { ConsoleCorePageHead, Topbar } from "../../../components";
+import { ConsoleCorePageHead } from "../../../components";
 import { NextPageWithLayout } from "../../_app";
 import { useAccessToken } from "../../../lib/useAccessToken";
 import { useTrackToken } from "../../../lib/useTrackToken";
