+++ apps/console/src/pages/[entity]/dashboard/pipeline/[id].tsx
@@ -4,9 +4,10 @@ import { Logo } from "@instill-ai/design-system";
 import {
   DashboardPipelineDetailsPageMainView,
   PageBase,
+  Topbar,
 } from "@instill-ai/toolkit";
 
-import { ConsoleCorePageHead, Topbar } from "../../../../components";
+import { ConsoleCorePageHead } from "../../../../components";
 import { NextPageWithLayout } from "../../../_app";
 import { useAccessToken } from "../../../../lib/useAccessToken";
 import { useTrackToken } from "../../../../lib/useTrackToken";
