+++ apps/console/src/pages/[entity]/models/[id].tsx
@@ -1,9 +1,13 @@
 import * as React from "react";
 import { useRouter } from "next/router";
-import { ModelHubSettingPageMainView, PageBase } from "@instill-ai/toolkit";
+import {
+  ModelHubSettingPageMainView,
+  PageBase,
+  Topbar,
+} from "@instill-ai/toolkit";
 import { Logo } from "@instill-ai/design-system";
 
-import { ConsoleCorePageHead, Topbar } from "../../../components";
+import { ConsoleCorePageHead } from "../../../components";
 import { NextPageWithLayout } from "../../_app";
 import { useAccessToken } from "../../../lib/useAccessToken";
 import { useTrackToken } from "../../../lib/useTrackToken";
