+++ apps/console/src/pages/[entity]/models/create.tsx
@@ -2,9 +2,14 @@ import * as React from "react";
 import { GetServerSideProps } from "next";
 import { useRouter } from "next/router";
 import { Logo } from "@instill-ai/design-system";
-import { env, ModelHubCreatePageMainView, PageBase } from "@instill-ai/toolkit";
+import {
+  env,
+  ModelHubCreatePageMainView,
+  PageBase,
+  Topbar,
+} from "@instill-ai/toolkit";
 
-import { ConsoleCorePageHead, Topbar } from "../../../components";
+import { ConsoleCorePageHead } from "../../../components";
 import { NextPageWithLayout } from "../../_app";
 import { useAccessToken } from "../../../lib/useAccessToken";
 import { useTrackToken } from "../../../lib/useTrackToken";
