+++ apps/console/src/pages/[entity]/pipelines/[id]/index.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import { PageBase, ViewPipeline } from "@instill-ai/toolkit";
+import { PageBase, ViewPipeline, Topbar } from "@instill-ai/toolkit";
 import { Logo } from "@instill-ai/design-system";
 
-import { ConsoleCorePageHead, Topbar } from "../../../../components";
+import { ConsoleCorePageHead } from "../../../../components";
 import { NextPageWithLayout } from "../../../_app";
 import { useAccessToken } from "../../../../lib/useAccessToken";
 import { useTrackToken } from "../../../../lib/useTrackToken";
