+++ apps/console/src/pages/[entity]/pipelines/index.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import { PageBase, ViewPipelines } from "@instill-ai/toolkit";
+import { PageBase, ViewPipelines, Topbar } from "@instill-ai/toolkit";
 import { Logo } from "@instill-ai/design-system";
 
-import { ConsoleCorePageHead, Topbar } from "../../../components";
+import { ConsoleCorePageHead } from "../../../components";
 import { NextPageWithLayout } from "../../_app";
 import { useAccessToken } from "../../../lib/useAccessToken";
 import { useTrackToken } from "../../../lib/useTrackToken";
