+++ apps/console/src/pages/onboarding.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { Logo } from "@instill-ai/design-system";
-import { PageBase, PageTitle } from "@instill-ai/toolkit";
+import { PageBase, PageTitle, Topbar } from "@instill-ai/toolkit";
 
 import { NextPageWithLayout } from "./_app";
-import { OnboardingForm, Topbar, ConsoleCorePageHead } from "../components";
+import { OnboardingForm, ConsoleCorePageHead } from "../components";
 
 const OnBoardingPage: NextPageWithLayout = () => {
   return (
