+++ packages/toolkit/src/view/pipeline-builder/Flow.tsx
@@ -58,8 +58,6 @@ export type FlowProps = {
   setReactFlowInstance: React.Dispatch<
     React.SetStateAction<Nullable<ReactFlowInstance>>
   >;
-  accessToken: Nullable<string>;
-  enableQuery: boolean;
   isLoading: boolean;
   isError: boolean;
   appEnv: InstillAppEnv;
@@ -67,14 +65,8 @@ export type FlowProps = {
 
 export const Flow = React.forwardRef<HTMLDivElement, FlowProps>(
   (props, ref) => {
-    const {
-      reactFlowInstance,
-      setReactFlowInstance,
-      accessToken,
-      enableQuery,
-      isLoading,
-      appEnv,
-    } = props;
+    const { reactFlowInstance, setReactFlowInstance, isLoading, appEnv } =
+      props;
 
     const {
       nodes,
@@ -90,10 +82,7 @@ export const Flow = React.forwardRef<HTMLDivElement, FlowProps>(
 
     return (
       <div className="relative flex flex-1 flex-col">
-        <BackToLatestVersionTopBar
-          enableQuery={enableQuery}
-          accessToken={accessToken}
-        />
+        <BackToLatestVersionTopBar />
         <div className="relative flex h-full w-full flex-1">
           <div ref={ref} className="h-full w-full flex-1">
             <ReactFlow
